requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.0
openai>=1.10.0,<2.0.0
anthropic>=0.16.0,<1.0.0
numpy==1.24.3
scipy==1.11.4
matplotlib==3.7.2
pandas==2.0.3
python-dotenv==1.0.0
pydantic==2.5.0
rich==13.7.0
textblob==0.17.1 